# Conflicts:
#	src/main/java/com/deepsoft/haolifa/service/impl/OrderProductServiceImpl.java

+++ src/main/java/com/deepsoft/haolifa/model/dto/export/DemandAmountDto.java
@@ -3,6 +3,8 @@ package com.deepsoft.haolifa.model.dto.export;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * 零星物料出，入库实体
  *
@@ -18,15 +20,15 @@ public class DemandAmountDto {
     private String demandName;
 
     @ApiModelProperty(value = "销售总金额")
-    private Double saleAmount;
+    private BigDecimal saleAmount;
 
     @ApiModelProperty(value = "开票总金额")
-    private Double invoiceAmount;
+    private BigDecimal invoiceAmount;
 
     @ApiModelProperty(value = "发货总金额")
-    private Double deliveryAmount;
+    private BigDecimal deliveryAmount;
 
     @ApiModelProperty(value = "回款总金额")
-    private Double refundAmount;
+    private BigDecimal refundAmount;
 
 }
